+++ example/lib/more_examples_section.dart
@@ -1,8 +1,12 @@
 import 'package:example/multiple_view_models_example/multiple_view_models_example.dart';
+import 'package:example/post_frame_callback_example/post_frame_callback_example.dart';
+import 'package:example/view_model_stateless_widget_example/view_model_stateless_example.dart';
 import 'package:flutter/material.dart';
 
 final _moreExamples = {
-  "Multiple ViewModels Example": const MultipleViewModelsExample()
+  "Multiple ViewModels Example": const MultipleViewModelsExample(),
+  "PostFrameCallback Example": const PostFrameCallbackExample(),
+  "ViewModelStatelessWidget Example": const ViewModelStatelessWidgetExample(),
 };
 
 class MoreExamplesSection extends StatelessWidget {